refactor: 총 거리 단위 변경 사항 반영(km -> m)

+++ src/main/java/wad/seoul_nolgoat/service/search/dto/DistanceSortCombinationDto.java
@@ -15,7 +15,7 @@ public class DistanceSortCombinationDto {
     private final int totalRounds;
 
     @Setter
-    private double totalDistance;
+    private int totalDistance;
 
     public DistanceSortCombinationDto(
             StoreForDistanceSortDto firstStore,
